✨ feat: 添加 result 缓存
